requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.31.0
PyMuPDF==1.23.8
python-docx==1.1.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
google-generativeai==0.3.2
python-dotenv==1.0.0
pydantic==2.5.0
numpy==1.24.3
aiofiles==0.24.0
httpx==0.25.2
mangum==0.17.0